Correct syntax of architectures value
Putting the library.properties architectures value in quotes causes it to be interpreted as a specific architecture named "*" instead of a wildcard match. This causes the following warning whenever the library is used:

WARNING: library StateMachine claims to run on ("*") architecture(s) and may be incompatible with your current board which runs on (avr) architecture(s).