Move example sketch to appropriately named folder
The Arduino IDE requires sketches to be located in a folder of the same name. If you open a sketch that is not in an appropriately named folder the Arduino IDE forces the creation of that folder before it will be opened. This problem caused the example sketch to not appear under File > Examples > StateMachine.